+++ examples/data-sources/iosxe_policy_map_event/data-source.tf
@@ -1,0 +1,4 @@
+data "iosxe_policy_map_event" "example" {
+  name       = "dot1x_policy"
+  event_type = "authentication-success"
+}