+++ examples/resources/iosxe_policy_map_event/import.sh
@@ -1,0 +1,1 @@
+terraform import iosxe_policy_map_event.example "Cisco-IOS-XE-native:native/policy/Cisco-IOS-XE-policy:policy-map=dot1x_policy/event=authentication-success"