+++ docs/guides/changelog.md
@@ -37,6 +37,7 @@ description: |-
 - Add `message_digest_keys` attribute to `iosxe_interface_ospf` resource and data source
 - Add `areas` and `passive_interface_default` attributes to `iosxe_ospf_vrf` resource and data source
 - Add `iosxe_policy_map` resource and data source
+- Add `iosxe_policy_map_event` resource and data source
 
 ## 0.3.3